fix(api): drop the int32 conversion that trips gosec G115
gosec fails the backend job on programs.go:241, `int32(plan.FailureCount)` —
G115 (CWE-190), integer overflow conversion int -> int32, HIGH severity. The
finding is on main, not on any open branch: it landed with the progression
preview in 4604542 and has been failing ever since.

It reached main because CI never re-ran gosec on that branch's final head. Run
#321 caught it. The next push to the branch touched only src/ui, so the `go`
job was path-filtered out, and go.yml's result gate counts a skipped build as a
pass ("backend checks OK"). The required check went green against a head whose
backend had never been linted, and PR #70 merged. Every main run since has been
UI-only, so it stayed hidden until the next backend PR ran gosec again.

Fix the conversion rather than suppress the warning: this repo has no #nosec in
first-party code, and the convention (sessions.go, api.go) is a provable bound.
Here the cleanest proof is not converting at all. progression.Plan.FailureCount
is an int, the DTO is hand-written and populated in exactly one place, and
openapi.yaml declares plain `type: integer` with no int32 format — so the int32
was arbitrary. FailuresBeforeDeload moves with it: same struct, same kind of
value, and it reads as arbitrary if only one of the pair changes.

The wire format does not change. encoding/json marshals int and int32
identically, verified byte-for-byte, so the OpenAPI contract, the generated
client and the UI (which types both as `number`) are untouched.

Note: gosec is not installed in the devcontainer sandbox, so this could not be
verified locally against gosec itself. The argument is structural — G115 fires
only on integer conversions, and there is no longer a conversion. go vet,
golangci-lint and go test -short all pass.

+++ src/api/internal/api/programs.go
@@ -238,7 +238,7 @@ func (s *Server) prescribe(ctx context.Context, programID, dayID int32) ([]presc
 			RestSeconds:  restSecondsDefault,
 			Progression: progressionInfoDTO{
 				Status:               string(plan.Status),
-				FailureCount:         int32(plan.FailureCount),
+				FailureCount:         plan.FailureCount,
 				FailuresBeforeDeload: progression.FailuresBeforeDeload,
 				PreviousWeightLb:     plan.PreviousLb,
 			},
